functions/requirements.txt
firebase_functions~=0.1.0
firebase_admin~=6.2.0
google-cloud-bigquery~=3.11.0
google-cloud-firestore~=2.11.0
google-cloud-logging~=3.8.0
requests~=2.32.3
